feat: Add SignUpPage css (#2)
