+++ src/pages/FormSuccess.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 const FormSuccess = () => {
   return (
-    <div>ㅎㅇㅎㅇㅎㅇ</div>
+    <div>회원가입이 완료되었습니다.</div>
   );
 };
