+++ src/pages/useForm.js
@@ -2,10 +2,7 @@ import {useState, useEffect} from 'react';
 
 const useForm = (callback, validate) => {
     const [values, setValues] = useState({
-        username: '',
-        email: '',
-        password: '',
-        password2: ''
+        userNickname: '',
     });
     const [errors, setErrors]  = useState({});
     const [isSubmitting, setIsSubmitting] = useState(false);
